+++ src/test/codemods/__fixtures__/convert-fn-expr-to-arrow-fn.input.js
@@ -1,0 +1,11 @@
+/*$ { fixture: 'should-trigger-when-no-this' } $*/
+
+const a = function() { /*# { pos: 14 } #*/
+    let foo = 'foo';
+};
+
+/*$ { fixture: 'should-trigger-when-bound-this' } $*/
+
+const a = function() { /*# { pos: 14 } #*/
+    this.bar = 'bar';
+}.bind(this);