+++ src/test/codemods/__fixtures__/convert-fn-expr-to-arrow-fn.output.js
@@ -1,0 +1,11 @@
+/*$ { fixture: 'should-trigger-when-no-this' } $*/
+
+const a = () => {
+    let foo = 'foo';
+};
+
+/*$ { fixture: 'should-trigger-when-bound-this' } $*/
+
+const a = () => {
+    this.bar = 'bar';
+};